github-monitor/requirements.txt
requests>=2.31.0
PyYAML>=6.0
schedule>=1.2.0
pytz>=2024.1
